+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Security
+
+- Bumpd dependencies versions.
+
 ## [1.4.0] - 2020-04-23
 
 ### Added